+++ layouts/default.vue
@@ -1,9 +1,15 @@
 <template>
-  <nuxt />
+  <div>
+    <sidemenu-menu />
+    <nuxt />
+  </div>
 </template>
 <script>
+import sidemenuMenu from '../components/sidemenu-menu';
+
 export default {
   components: {
+    sidemenuMenu,
   },
 };
 </script>