+++ nuxt.config.js
@@ -25,7 +25,7 @@ module.exports = {
   },
 
   router: {
-    middleware: ['auth'],
+    middleware: ['auth', 'menu'],
   },
 
   env: {
@@ -52,8 +52,12 @@ module.exports = {
           jwt: 'jwt',
           rides: 'rides',
           campuses: 'campuses',
+          users: 'users',
         },
         withAuth: true,
+        authPlugins: [
+          'auth-renew',
+        ],
       },
     ],
   ],
@@ -71,6 +75,8 @@ module.exports = {
           'faAngleDown',
           'faHeadset',
           'faSpinner',
+          'faBars',
+          'faCar',
         ],
       },
     ],