+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
-## Current
+## Version 1.3.0
+### Features
+* When the passenger is on board, show the arrival label in notification space.
+* Enable High Accuracy
 ## Version 1.2.0
 ### Features
 * Use common library